Crate ui

UI - terminal user interface to visualize tree exploration for PUCT

Usage: cargo run --release --bin ui -- -c breakthrough -m alpha

Keyboard and mouse can be used to play the game step by step while inspecting the tree search.

Structs

GameDuelUI
Interface for game matches.

GuiEventSender
Communication channel from Simulator to GUI.

TreeEntry
Entry for the tree view.

Enums

GuiToSimChannel
Communication type from GUI to Simulator.

Functions

event_loop_alpha
AlphaZero event loop, managing the game instance.

event_loop_muz
MuZero event loop, managing the game instance.

expand_tree
Expand the tree view by looking for the children and inserting new nodes.

main
Entry point.

run_cursive
Dispatch cursive instance according to the chosen method.

run_gym
Use MuZero with remote Gym.

Type Definitions

Result